Reorgansing "query" scripts in the "common" folder to make it easier to find files.
