--- common/deposits_trace_query.py
@@ -1,5 +1,0 @@
-from common.deposits_query_builder import DepositsQueryBuilder
-from common.trace_query import TraceQuery
-
-class DepositsTraceQuery(TraceQuery, DepositsQueryBuilder):
-    pass
