+++ common/es_queries/deposits_trace_query.py
@@ -1,0 +1,5 @@
+from common.es_queries.deposits_query_builder import DepositsQueryBuilder
+from common.es_queries.trace_query import TraceQuery
+
+class DepositsTraceQuery(TraceQuery, DepositsQueryBuilder):
+    pass
